requirements.txt
biopython==1.66
numpy==1.11.0
